Move MinimiseLength into optimiser module
