+++ tests/bluemira/builders/EUDEMO/first_wall/test_wall.py
@@ -28,6 +28,7 @@ import pytest
 
 from bluemira.builders.EUDEMO.first_wall import WallBuilder
 
+OPTIMISER_MODULE_REF = "bluemira.geometry.optimisation"
 WALL_MODULE_REF = "bluemira.builders.EUDEMO.first_wall.wall"
 
 CONFIG = {
@@ -42,7 +43,7 @@ CONFIG = {
     },
     "runmode": "mock",
     "name": "First Wall",
-    "problem_class": f"{WALL_MODULE_REF}::MinimiseLength",
+    "problem_class": f"{OPTIMISER_MODULE_REF}::MinimiseLength",
 }
 PARAMS = {
     "Name": "First Wall Example",
